Catch any exceptions with unarchiving accounts
If an account subclass doesn't exist on the device, an exception will be thrown. This can happen now because the accounts are stored in the keychain, which could be synced. So if you update an app on one device and not on another they could have mis-matched set of account classes.