Fix for loading local URLs
Local URLs return an NSURLResponse rather than an NSHTTPURLResponse, so we won't get a content type or status code.

Loading local URLs is useful for unit testing, so going through each method for getting the content object shouldn't be a problem in this case. If a server responds without a content type, this will handle it better, but performance could be slow due to trying some incorrect methods.